Initial implementation, continued (29):
- Arduino sketch size optimization:
  * The sketch now fits into Arduino Uno, but runs unstable
  * Implemented SERIAL_PRINTF & SERIAL_PRINTLNF macros, which
    automagically stores format string in program memory and
    retrieves it when message is being printed
  * Replaced all Serial.print/ln calls with SERIAL_* macros
  * Removed PSTR macro
  * Sketch now works with static IP, reducing program memory
    and SRAM usage even further
- tdsl::span, tdsl::binary_reader and tdsl::binary_writer now
  uses platform defined size variables instead of using *int
  [32|64]
  * It makes no sense to have >sizeof(max pmem) variables for
    addresses and size variables, so they're all dependent to
    the platform. tdsl::span<> is only 4 bytes in arduino AVR
    whereas it is 16 bytes in x86-64.
  * Replaced upper-cast arithmetics with boundary checks
- type_traits: added make_unsigned, is_floating_point, is_arithmetic
  is_signed
- inttypes: added ssize_t, numeric_limits::min_value, numeric_
  limits::max_value
- vscode: added code workspace file to open project faster in devenv

+++ tdslite/include/tdslite/util/tdsl_span.hpp
@@ -144,8 +144,8 @@ namespace tdsl {
 
         using element_type               = T;
         using value_type                 = typename traits::remove_cv<T>::type;
-        using size_type                  = tdsl::uint32_t;
-        using ssize_type                 = tdsl::int32_t;
+        using size_type                  = tdsl::size_t;
+        using ssize_type                 = tdsl::ssize_t;
         using pointer                    = T *;
         using const_pointer              = const T *;
         using reference                  = T &;
